Address integer overflows across guardian
Many ignores to avoid changing public interfaces and work through issues
caused by libcontainer using int, but garden + linux using uint32 for
things.

Updated a few things for better types to avoid casting, or casting in
the opposite direction for more safety

+++ rundmc/execrunner/dadoo/winsize.go
@@ -26,7 +26,7 @@ func SetWinSize(f *os.File, ws garden.WindowSize) error {
 		uintptr(unsafe.Pointer(&struct {
 			Rows uint16
 			Cols uint16
-		}{Rows: uint16(ws.Rows), Cols: uint16(ws.Columns)})),
+		}{Rows: ws.Rows, Cols: ws.Columns})),
 		0, 0, 0,
 	)
